mudar a responsa do banco

+++ jogo/sair.js
@@ -1,11 +1,22 @@
+const flag = localStorage.getItem("flag")
+
 export function sair() {
     const sair = document.querySelector("[sair]")
 
     sair.addEventListener("click", function () {
-        console.log("sair")
-        localStorage.clear()
-        location.href = '../index.html'
-    })
-}
 
+        if (flag == 1) {
+            alert("você não pode sair, para de ser gayyyy")
+        } else {
+            console.log("sair")
+            localStorage.clear()
+            location.href = '../index.html'
+        }
+       
+
+      
+    })
+
+  
+}
